ci: Use `ubuntu-latest` for Py27
Py27 is build now (or cached), so we can do that on the `ubuntu-latest`
image. Py35, and Py36 are still available on `ubuntu-20.04`, and we use
those and don't need to build them too (yet).

Keeps those older versions separate to allow for failures when testing
since they are not crucial.