pymc: Fix interpolation
Fixes interpolation edges in the PyMC3 and PyMC4/5 implementations by
using the actual `x` values to determine out of bounds locations instead
of relying on the index. The index is fixed to 0 or N and does not tell
if the actual value is smaller or larger. Also limits the index to the
maximum index if there is no interpolation requested, to handle the
right end point. Limits the index to last - 1 for interpolation, since
we use that and the one before. Otherwise there will be an out-of-bounds
access to the array.